+++ packages/core/src/index.ts
@@ -4,3 +4,4 @@ export * from "./session.js";
 export * from "./stats.js";
 export * from "./unlock-pool.js";
 export * from "./points.js";
+export * from "./tank-themes.js";